JCF: DUNE-DAQ/daq-deliverables#73: merge the submodules in light of the detdataformats recent split
