+++ pybindsrc/daqethheader.cpp
@@ -6,7 +6,6 @@
  * received with this code.
  */
 
-#include "submodules.hpp"
 #include "detdataformats/DAQEthHeader.hpp"
 
 #include <pybind11/operators.h>
@@ -22,11 +21,6 @@ namespace dunedaq::detdataformats::python {
 void register_daqethheader(py::module& m) {
 
   py::class_<DAQEthHeader>(m, "DAQEthHeader", py::buffer_protocol())
-    // .def(py::init())
-    // .def(py::init([](py::capsule capsule) {
-    //     auto dhp = *static_cast<DAQEthHeader*>(capsule.get_pointer());
-    //     return dhp;
-    // } ))
     .def_property("version", 
       [](DAQEthHeader& self) -> uint32_t { return self.version; }, 
       [](DAQEthHeader& self, uint32_t version) { self.version = version; }
