--- python/detdataformats/hsi/__init__.py
@@ -1,1 +1,0 @@
-from .._daq_detdataformats_py.hsi import *
